chore: Treate the prof=true backtrace method as an enum
When --enable-prof is set during configure, detection of a suitable backtrace
method is performed stopping with the first supported solution. This can be
augmented by explicitly enabling or disabling detection of specific methods.
To better model this in a bazel context, the //settings/flags:enable_prof is
extended to have enumerable values where the backtrace method can be explicitly
declared.